+++ bootique-rabbitmq-client/src/main/java/io/bootique/rabbitmq/client/pubsub/RmqSubEndpointFactory.java
@@ -20,7 +20,7 @@ package io.bootique.rabbitmq.client.pubsub;
 
 import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
-import io.bootique.rabbitmq.client.connection.RmqConnectionManager;
+import io.bootique.rabbitmq.client.channel.RmqChannelManager;
 import io.bootique.rabbitmq.client.topology.RmqTopologyManager;
 import io.bootique.shutdown.ShutdownManager;
 
@@ -38,9 +38,9 @@ public class RmqSubEndpointFactory {
     private String routingKey;
     private boolean autoAck = true;
 
-    public RmqSubEndpoint create(RmqConnectionManager connectionManager, RmqTopologyManager topologyManager, ShutdownManager shutdownManager) {
+    public RmqSubEndpoint create(RmqChannelManager channelManager, RmqTopologyManager topologyManager, ShutdownManager shutdownManager) {
         Objects.requireNonNull(connection, "Subscriber connection name is undefined");
-        RmqEndpointDriver driver = new RmqEndpointDriver(connectionManager, topologyManager, connection);
+        RmqEndpointDriver driver = new RmqEndpointDriver(channelManager, topologyManager, connection);
 
         RmqSubEndpoint endpoint = new RmqSubEndpoint(driver, exchange, queue, routingKey, autoAck);
         shutdownManager.addShutdownHook(() -> endpoint.close());